Fix memory leaks on Stream
Fixed some memory leaks on the stream caused by several factors including bound
DOM events that were never released. There are still leaks, but many fewer.

Some components of StreamActions were refactored into components for
clarity.

The DatePicker was a large source of memory leaks and was non-functional, so we
commented it out.
